Rename patches module.
