+++ emle/models/_emle.py
@@ -36,19 +36,19 @@ import torchani as _torchani
 from torch import Tensor
 from typing import Optional, Tuple, List
 
-from . import _torchani_patches
+from . import _patches
 
 # Monkey-patch the TorchANI BuiltInModel and BuiltinEnsemble classes so that
 # they call self.aev_computer using args only to allow forward hooks to work
 # with TorchScript.
-_torchani.models.BuiltinModel = _torchani_patches.BuiltinModel
-_torchani.models.BuiltinEnsemble = _torchani_patches.BuiltinEnsemble
+_torchani.models.BuiltinModel = _patches.BuiltinModel
+_torchani.models.BuiltinEnsemble = _patches.BuiltinEnsemble
 
 try:
     import NNPOps as _NNPOps
     import NNPOps.neighbors.getNeighborPairs as _getNeighborPairs
 
-    _NNPOps.OptimizedTorchANI = _torchani_patches.OptimizedTorchANI
+    _NNPOps.OptimizedTorchANI = _patches.OptimizedTorchANI
 
     _has_nnpops = True
 except:
